// src/models/V1Pod.ts
// tslint:disable
/**
 * Kubernetes
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.14.11
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { KubeObject } from '../runtime'
import V1ObjectMeta from './V1ObjectMeta'
import V1PodSpec from './V1PodSpec'
import V1PodStatus from './V1PodStatus'

/**
 * Pod is a collection of containers that can run on a host. This resource is created by clients and scheduled onto hosts.
 * @export
 * @interface V1Pod
 */
export default interface V1Pod extends KubeObject {
  /**
   * APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources
   * @type {string}
   * @memberof V1Pod
   */
  apiVersion?: string

  /**
   * Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds
   * @type {string}
   * @memberof V1Pod
   */
  kind?: string

  /**
   *
   * @type {V1ObjectMeta}
   * @memberof V1Pod
   */
  metadata?: V1ObjectMeta

  /**
   *
   * @type {V1PodSpec}
   * @memberof V1Pod
   */
  spec?: V1PodSpec

  /**
   *
   * @type {V1PodStatus}
   * @memberof V1Pod
   */
  status?: V1PodStatus
}