// src/models/V1Namespace.ts
// tslint:disable
/**
 * Kubernetes
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.14.11
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { KubeObject } from '../runtime'
import V1NamespaceSpec from './V1NamespaceSpec'
import V1NamespaceStatus from './V1NamespaceStatus'
import V1ObjectMeta from './V1ObjectMeta'

/**
 * Namespace provides a scope for Names. Use of multiple namespaces is optional.
 * @export
 * @interface V1Namespace
 */
export default interface V1Namespace extends KubeObject {
  /**
   * APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#resources
   * @type {string}
   * @memberof V1Namespace
   */
  apiVersion?: string

  /**
   * Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/api-conventions.md#types-kinds
   * @type {string}
   * @memberof V1Namespace
   */
  kind?: string

  /**
   *
   * @type {V1ObjectMeta}
   * @memberof V1Namespace
   */
  metadata?: V1ObjectMeta

  /**
   *
   * @type {V1NamespaceSpec}
   * @memberof V1Namespace
   */
  spec?: V1NamespaceSpec

  /**
   *
   * @type {V1NamespaceStatus}
   * @memberof V1Namespace
   */
  status?: V1NamespaceStatus
}